+++ core/src/main/java/br/com/jpcchaves/core/domain/model/Picture.java
@@ -1,4 +1,4 @@
-package br.com.jpcchaves.core.domain;
+package br.com.jpcchaves.core.domain.model;
 
 import java.util.Objects;
